fix(matching): show human-readable interest labels in candidate views
Extract INTERESTS lookup into shared util and apply interestLabel()
in candidate-card, match-card, and partner profile. Unifies labels
between onboarding modal and profile screen.

Bug: #317
