+++ apps/native/components/onboarding-modal.tsx
@@ -19,6 +19,7 @@ import { LanguagePickerModal } from "@/components/language-picker-modal";
 import { authClient } from "@/lib/auth-client";
 import { queryClient, trpc } from "@/utils/trpc";
 import { pickAndEncodeProfilePicture } from "@/utils/profile-picture";
+import { INTERESTS } from "@/utils/interest-labels";
 
 const GOLD = "#F2C94C";
 const MUTED_BORDER = "#D9C9BC";
@@ -40,27 +41,6 @@ const LEVEL_BLOCKS = [
   { value: "advanced" as LearningProficiency, label: "C1–C2", sub: "Advanced" },
 ];
 
-const INTERESTS: { value: InterestValue; label: string }[] = [
-  { value: "modern_art", label: "Art" },
-  { value: "tech_coding", label: "Tech" },
-  { value: "jazz_music", label: "Music" },
-  { value: "culinary_arts", label: "Cooking" },
-  { value: "sustainability", label: "Sustainability" },
-  { value: "cinephile", label: "Film" },
-  { value: "cosmology", label: "Cosmology" },
-  { value: "photography", label: "Photography" },
-  { value: "board_games", label: "Board games" },
-  { value: "hiking_outdoors", label: "Hiking" },
-  { value: "yoga_wellness", label: "Yoga" },
-  { value: "literature", label: "Literature" },
-  { value: "entrepreneurship", label: "Startups" },
-  { value: "design_architecture", label: "Design" },
-  { value: "travel", label: "Travel" },
-  { value: "gaming", label: "Gaming" },
-  { value: "fitness_sports", label: "Football" },
-  { value: "philosophy", label: "Philosophy" },
-  { value: "theatre", label: "Theatre" },
-];
 
 const LANGUAGE_FLAGS: Record<string, string> = {
   Afrikaans: "🇿🇦", Albanian: "🇦🇱", Arabic: "🇸🇦", Armenian: "🇦🇲",
